backend/requirements.txt
django==5.2
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.4.0
django-filter==24.3
django-unfold==0.80.2
psycopg2-binary==2.9.9
stripe==10.12.0
paypalrestsdk==1.13.3
Pillow==10.4.0
gunicorn==22.0.0
